Fixes https://github.com/allenai/tools/issues/18
This upgrades the dependencies to the code in Typesafe's sbt-web
family of plugins. There have been several bug fixes and changes
to the underlying sbt-web plugin that improve stability and performance,
including a fix for asset management that fixes the issue we've had
with HTML file changes not being picked up right away.