Publish sbt-travis-publisher to sonatype
This plugin is used by `common` project, which is going to become
public. As such, we need to remove dependencies on our private
plugins. In the long run, maybe it makes sense to make all of our
plugins public. For now, this just makes sbt-travis-publisher public.

Note: when I signed up for the Sonatype account, I totally forgot to use
the ari-role username / email. I'll put an issue in to try and get that
fixed (either changing username on account or transfering ownership).